+++ hq/src/app/hq/layout.tsx
@@ -41,6 +41,9 @@ export default async function HqLayout({
         <Link className="button button-quiet button-small" href="/hq">
           💡 Ideas
         </Link>
+        <Link className="button button-quiet button-small" href="/hq/vote">
+          🗳️ Vote
+        </Link>
         <Link className="button button-quiet button-small" href="/hq/timeline">
           🗓️ Timeline
         </Link>